+++ src/Smartstore/Net/Mail/MailAttachment.cs
@@ -1,4 +1,5 @@
-﻿using Smartstore.IO;
+﻿using MimeKit.Utils;
+using Smartstore.IO;
 
 namespace Smartstore.Net.Mail
 {
@@ -15,6 +16,9 @@ namespace Smartstore.Net.Mail
 
     public class MailAttachment : Disposable
     {
+        private bool _isEmbedded;
+        private string _contentId;
+        
         public MailAttachment(IFile file)
         {
             Guard.NotNull(file, nameof(file));
@@ -59,6 +63,39 @@ namespace Smartstore.Net.Mail
         public DateTimeOffset? CreationDate { get; set; }
         public DateTimeOffset? ModificationDate { get; set; }
         public DateTimeOffset? ReadDate { get; set; }
+
+        /// <summary>
+        /// If true, embeds the attachment to the message body.
+        /// </summary>
+        public bool IsEmbedded
+        {
+            get => _isEmbedded;
+            set 
+            {
+                if (!value)
+                {
+                    _contentId = null;
+                }
+
+                _isEmbedded = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets the unique content id of the embedded attachment.
+        /// Use this id as reference in HTML body, e.g. $"&lt;img src='cid:{contentId}' /&gt;"
+        /// </summary>
+        /// <exception cref="InvalidOperationException">Thrown if attachment is not embedded.</exception>
+        /// <returns>Content id</returns>
+        public string GetContentId()
+        {
+            if (!_isEmbedded)
+            {
+                throw new InvalidOperationException("Content id can only be generated for embedded attachments.");
+            }
+
+            return _contentId ??= MimeUtils.GenerateMessageId();
+        }
 
         protected override void OnDispose(bool disposing)
         {